requirements.txt
flask
flask-cors
openai>=1.0.0
numpy
python-dotenv
